+++ src/core/f-series.c
@@ -45,7 +45,7 @@
 //
 REBOOL Series_Common_Action_Returns(
     REB_R *r, // `r_out` would be slightly confusing, considering R_OUT
-    struct Reb_Frame *frame_,
+    REBFRM *frame_,
     REBSYM action
 ) {
     REBVAL *value = D_ARG(1);
