+++ src/core/p-clipboard.c
@@ -34,7 +34,7 @@
 //
 //  Clipboard_Actor: C
 //
-static REB_R Clipboard_Actor(struct Reb_Frame *frame_, REBCTX *port, REBSYM action)
+static REB_R Clipboard_Actor(REBFRM *frame_, REBCTX *port, REBSYM action)
 {
     REBREQ *req;
     REBINT result;
