+++ src/core/p-serial.c
@@ -37,7 +37,7 @@
 //
 //  Serial_Actor: C
 //
-static REB_R Serial_Actor(struct Reb_Frame *frame_, REBCTX *port, REBSYM action)
+static REB_R Serial_Actor(REBFRM *frame_, REBCTX *port, REBSYM action)
 {
     REBREQ *req;    // IO request
     REBVAL *spec;   // port spec
