+++ src/core/p-signal.c
@@ -152,7 +152,7 @@ static int sig_word_num(REBSTR *canon)
 //
 //  Signal_Actor: C
 //
-static REB_R Signal_Actor(struct Reb_Frame *frame_, REBCTX *port, REBSYM action)
+static REB_R Signal_Actor(REBFRM *frame_, REBCTX *port, REBSYM action)
 {
     REBREQ *req;
     REBINT result;
