+++ src/core/p-timer.c
@@ -47,7 +47,7 @@
 //
 //  Event_Actor: C
 //
-static REB_R Event_Actor(struct Reb_Frame *frame_, REBCTX *port, REBCNT action)
+static REB_R Event_Actor(REBFRM *frame_, REBCTX *port, REBCNT action)
 {
     REBVAL *spec;
     REBVAL *state;
