+++ src/include/sys-action.h
@@ -172,17 +172,17 @@ inline static REBOOL SAME_SYM_NONZERO(REBSYM a, REBSYM b) {
 }
 
 // NATIVE! function
-typedef REB_R (*REBNAT)(struct Reb_Frame *frame_);
+typedef REB_R (*REBNAT)(REBFRM *frame_);
 #define REBNATIVE(n) \
-    REB_R N_##n(struct Reb_Frame *frame_)
+    REB_R N_##n(REBFRM *frame_)
 
 // ACTION! function (one per each DATATYPE!)
-typedef REB_R (*REBACT)(struct Reb_Frame *frame_, REBSYM a);
+typedef REB_R (*REBACT)(REBFRM *frame_, REBSYM a);
 #define REBTYPE(n) \
-    REB_R T_##n(struct Reb_Frame *frame_, REBSYM action)
+    REB_R T_##n(REBFRM *frame_, REBSYM action)
 
 // PORT!-action function
-typedef REB_R (*REBPAF)(struct Reb_Frame *frame_, REBCTX *p, REBSYM a);
+typedef REB_R (*REBPAF)(REBFRM *frame_, REBCTX *p, REBSYM a);
 
 // COMMAND! function
 typedef REB_R (*CMD_FUNC)(REBCNT n, REBSER *args);
