+++ src/include/sys-do.h
@@ -82,52 +82,6 @@ inline static REBOOL IS_QUOTABLY_SOFT(const RELVAL *v) {
 }
 
 
-inline static REBOOL Is_Any_Function_Frame(struct Reb_Frame *f) {
-    return LOGICAL(f->eval_type == ET_FUNCTION || f->eval_type == ET_LOOKBACK);
-}
-
-// While a function frame is fulfilling its arguments, the `f->param` will
-// be pointing to a typeset.  The invariant that is maintained is that
-// `f->param` will *not* be a typeset when the function is actually in the
-// process of running.  (So no need to set/clear/test another "mode".)
-//
-inline static REBOOL Is_Function_Frame_Fulfilling(struct Reb_Frame *f)
-{
-    assert(Is_Any_Function_Frame(f));
-    return NOT_END(f->param);
-}
-
-
-// It's helpful when looking in the debugger to be able to look at a frame
-// and see a cached string for the function it's running (if there is one).
-// The release build only considers the frame symbol valid if ET_FUNCTION
-//
-inline static void SET_FRAME_LABEL(struct Reb_Frame *f, REBSTR *label) {
-    assert(Is_Any_Function_Frame(f));
-    f->label = label;
-#if !defined(NDEBUG)
-    f->label_debug = cast(const char*, STR_HEAD(label));
-#endif
-}
-
-inline static void CLEAR_FRAME_LABEL(struct Reb_Frame *f) {
-#if !defined(NDEBUG)
-    f->label = NULL;
-    f->label_debug = NULL;
-#endif
-}
-
-inline static void SET_FRAME_VALUE(struct Reb_Frame *f, const RELVAL *value) {
-    f->value = value;
-
-#if !defined(NDEBUG)
-    if (NOT_END(f->value))
-        f->value_type = VAL_TYPE(f->value);
-    else
-        f->value_type = REB_MAX;
-#endif
-}
-
 //=////////////////////////////////////////////////////////////////////////=//
 //
 //  DO's LOWEST-LEVEL EVALUATOR HOOKING
@@ -185,7 +139,7 @@ inline static void SET_FRAME_VALUE(struct Reb_Frame *f, const RELVAL *value) {
 // optimize performance by working with the evaluator directly.
 //
 
-inline static void PUSH_CALL(struct Reb_Frame *f)
+inline static void PUSH_CALL(REBFRM *f)
 {
     f->prior = TG_Frame_Stack;
     TG_Frame_Stack = f;
@@ -196,12 +150,12 @@ inline static void PUSH_CALL(struct Reb_Frame *f)
         }
 }
 
-inline static void UPDATE_EXPRESSION_START(struct Reb_Frame *f) {
+inline static void UPDATE_EXPRESSION_START(REBFRM *f) {
     assert(NOT(f->flags & DO_FLAG_VA_LIST));
     f->expr_index = f->index;
 }
 
-inline static void DROP_CALL(struct Reb_Frame *f) {
+inline static void DROP_CALL(REBFRM *f) {
     if (f->flags & DO_FLAG_TOOK_FRAME_LOCK) {
         assert(GET_ARR_FLAG(f->source.array, SERIES_FLAG_LOCKED));
         CLEAR_ARR_FLAG(f->source.array, SERIES_FLAG_LOCKED);
@@ -223,10 +177,10 @@ inline static void DROP_CALL(struct Reb_Frame *f) {
 // is to hold a temporary lock on the array.
 //
 
-typedef struct Reb_Frame Reb_Enumerator;
+typedef REBFRM Reb_Enumerator;
 
 inline static void PUSH_SAFE_ENUMERATOR(
-    struct Reb_Frame *f,
+    REBFRM *f,
     const REBVAL *v
 ) {
     SET_FRAME_VALUE(f, VAL_ARRAY_AT(v));
@@ -265,7 +219,7 @@ inline static void PUSH_SAFE_ENUMERATOR(
 // a "virtual" next value (e.g. an old value saved by EVAL) are both indicated
 // by f->pending, hence a NULL test of that can be executed quickly.
 //
-inline static void FETCH_NEXT_ONLY_MAYBE_END(struct Reb_Frame *f) {
+inline static void FETCH_NEXT_ONLY_MAYBE_END(REBFRM *f) {
     TRACE_FETCH_DEBUG("FETCH_NEXT_ONLY_MAYBE_END", f, FALSE);
 
     assert(NOT_END(f->value));
@@ -306,7 +260,7 @@ inline static void FETCH_NEXT_ONLY_MAYBE_END(struct Reb_Frame *f) {
 //
 inline static void Try_Lookback_In_Prior_Frame(
     REBVAL *out,
-    struct Reb_Frame *prior
+    REBFRM *prior
 ){
     switch (prior->eval_type) {
     case ET_SET_WORD:
@@ -361,13 +315,13 @@ inline static void Try_Lookback_In_Prior_Frame(
 //
 inline static void DO_NEXT_REFETCH_MAY_THROW(
     REBVAL *out,
-    struct Reb_Frame *parent,
+    REBFRM *parent,
     REBUPT flags
 ){
     TRACE_FETCH_DEBUG("DO_NEXT_REFETCH_MAY_THROW", parent, FALSE);
 
-    struct Reb_Frame child_frame;
-    struct Reb_Frame *child = &child_frame;
+    REBFRM child_frame;
+    REBFRM *child = &child_frame;
 
     child->eval_type = Eval_Table[VAL_TYPE(parent->value)];
 
@@ -498,7 +452,7 @@ no_optimization:
 }
 
 
-inline static void QUOTE_NEXT_REFETCH(RELVAL *dest, struct Reb_Frame *f) {
+inline static void QUOTE_NEXT_REFETCH(RELVAL *dest, REBFRM *f) {
     TRACE_FETCH_DEBUG("QUOTE_NEXT_REFETCH", f, FALSE);
     COPY_VALUE(dest, f->value, f->specifier);
     CLEAR_VAL_FLAG(dest, VALUE_FLAG_EVALUATED);
@@ -542,8 +496,8 @@ inline static REBIXO DO_NEXT_MAY_THROW(
     REBCNT index,
     REBCTX *specifier
 ){
-    struct Reb_Frame frame;
-    struct Reb_Frame *f = &frame;
+    REBFRM frame;
+    REBFRM *f = &frame;
 
     SET_FRAME_VALUE(f, ARR_AT(array, index));
     if (IS_END(f->value)) {
@@ -570,6 +524,383 @@ inline static REBIXO DO_NEXT_MAY_THROW(
     assert(f->index > 1);
     return f->index - 1;
 }
+
+
+// Most common case of evaluator invocation in Rebol: the data lives in an
+// array series.  Generic routine takes flags and may act as either a DO
+// or a DO/NEXT at the position given.  Option to provide an element that
+// may not be resident in the array to kick off the execution.
+//
+inline static REBIXO Do_Array_At_Core(
+    REBVAL *out,
+    const RELVAL *opt_first, // must also be relative to specifier if relative
+    REBARR *array,
+    REBCNT index,
+    REBCTX *specifier,
+    REBFLGS flags
+) {
+    REBFRM f;
+
+    if (opt_first) {
+        SET_FRAME_VALUE(&f, opt_first);
+        f.index = index;
+    }
+    else {
+        // Do_Core() requires caller pre-seed first value, always
+        //
+        SET_FRAME_VALUE(&f, ARR_AT(array, index));
+        f.index = index + 1;
+    }
+
+    if (IS_END(f.value)) {
+        SET_VOID(out);
+        return END_FLAG;
+    }
+
+    f.out = out;
+    f.source.array = array;
+    f.specifier = specifier;
+    f.flags = flags;
+    f.gotten = NULL; // so ET_WORD and ET_GET_WORD do their own Get_Var
+    f.pending = NULL;
+
+    f.eval_type = Eval_Table[VAL_TYPE(f.value)];
+
+    Do_Core(&f);
+
+    if (THROWN(f.out))
+        return THROWN_FLAG; // !!! prohibits recovery from exits
+
+    return IS_END(f.value) ? END_FLAG : f.index;
+}
+
+
+// !!! Not yet implemented--concept is to accept a REBVAL[] array, rather
+// than a REBARR of values.
+//
+// !!! Considerations of this core interface are to see the values as being
+// potentially in non-contiguous points in memory, and advanced with some
+// skip length between them.  Additionally the idea of some kind of special
+// Rebol value or "REB_INSTRUCTION" to say how far to skip is a possibility,
+// which would be more general in the sense that it would allow the skip
+// distances to be generalized, though this would cost a pointer size
+// entity at each point.  The advantage of REB_INSTRUCTION is that only the
+// clients using the esoteric ability would be paying anything for it or
+// the API complexity, but if an important client like Ren-C++ it might
+// be worth the savings.
+//
+// Note: Functionally it would be possible to assume a 0 index and require
+// the caller to bump the value pointer as necessary.  But an index-based
+// interface is likely useful to avoid the bookkeeping required for the caller.
+//
+inline static REBIXO Do_Values_At_Core(
+    REBVAL *out,
+    REBFLGS flags,
+    const REBVAL *opt_head,
+    const REBVAL values[],
+    REBCNT index
+) {
+    fail (Error(RE_MISC));
+}
+
+
+//
+//  Reify_Va_To_Array_In_Frame: C
+//
+// For performance and memory usage reasons, a variadic C function call that
+// wants to invoke the evaluator with just a comma-delimited list of REBVAL*
+// does not need to make a series to hold them.  Do_Core is written to use
+// the va_list traversal as an alternate to DO-ing an ARRAY.
+//
+// However, va_lists cannot be backtracked once advanced.  So in a debug mode
+// it can be helpful to turn all the va_lists into arrays before running
+// them, so stack frames can be inspected more meaningfully--both for upcoming
+// evaluations and those already past.
+//
+// A non-debug reason to reify a va_list into an array is if the garbage
+// collector needs to see the upcoming values to protect them from GC.  In
+// this case it only needs to protect those values that have not yet been
+// consumed.
+//
+// Because items may well have already been consumed from the va_list() that
+// can't be gotten back, we put in a marker to help hint at the truncation
+// (unless told that it's not truncated, e.g. a debug mode that calls it
+// before any items are consumed).
+//
+inline static void Reify_Va_To_Array_In_Frame(
+    REBFRM *f,
+    REBOOL truncated
+) {
+    REBDSP dsp_orig = DSP;
+
+    assert(f->flags & DO_FLAG_VA_LIST);
+
+    if (truncated) {
+        REBVAL temp;
+        Val_Init_Word(&temp, REB_WORD, Canon(SYM___OPTIMIZED_OUT__));
+
+        DS_PUSH(&temp);
+    }
+
+    if (NOT_END(f->value)) {
+        do {
+            DS_PUSH_RELVAL_MAYBE_VOID(f->value, f->specifier);
+            FETCH_NEXT_ONLY_MAYBE_END(f);
+        } while (NOT_END(f->value));
+
+        if (truncated)
+            f->index = 2; // skip the --optimized-out--
+        else
+            f->index = 1; // position at the start of the extracted values
+    }
+    else {
+        // Leave at the END, but give back the array to serve as
+        // notice of the truncation (if it was truncated)
+        //
+        f->index = 0;
+    }
+
+    if (DSP != dsp_orig) {
+        f->source.array = Pop_Stack_Values(dsp_orig);
+        MANAGE_ARRAY(f->source.array); // held alive while frame running
+
+        SET_ARR_FLAG(f->source.array, SERIES_FLAG_LOCKED);
+        SET_ARR_FLAG(f->source.array, ARRAY_FLAG_VOIDS_LEGAL);
+        f->flags |= DO_FLAG_TOOK_FRAME_LOCK;
+    }
+    else {
+        // The series needs to be locked during Do_Core, but it doesn't have
+        // to be unique.  Use empty array but don't say we locked it.
+
+        assert(GET_ARR_FLAG(EMPTY_ARRAY, SERIES_FLAG_LOCKED));
+        f->source.array = EMPTY_ARRAY;
+    }
+
+    if (truncated)
+        SET_FRAME_VALUE(f, ARR_AT(f->source.array, 1)); // skip `--optimized--`
+    else
+        SET_FRAME_VALUE(f, ARR_HEAD(f->source.array));
+
+    // We clear the DO_FLAG_VA_LIST, assuming that the truncation marker is
+    // enough information to record the fact that it was a va_list (revisit
+    // if there's another reason to know what it was...)
+
+    f->flags &= ~DO_FLAG_VA_LIST;
+
+    assert(f->pending == VA_LIST_PENDING);
+    f->pending = NULL;
+}
+
+
+// (va_list by pointer: http://stackoverflow.com/a/3369762/211160)
+//
+// Central routine for doing an evaluation of an array of values by calling
+// a C function with those parameters (e.g. supplied as arguments, separated
+// by commas).  Uses same method to do so as functions like printf() do.
+//
+// In R3-Alpha this style of invocation was specifically used to call single
+// Rebol functions.  It would use a list of REBVAL*s--each of which could
+// come from disjoint memory locations and be passed directly with no
+// evaluation.  Ren-C replaced this entirely by adapting the evaluator to
+// use va_arg() lists for the same behavior as a DO of an ARRAY.
+//
+// The previously accomplished style of execution with a function which may
+// not be in the arglist can be accomplished using `opt_first` to put that
+// function into the optional first position.  To instruct the evaluator not
+// to do any evaluation on the values supplied as arguments after that
+// (corresponding to R3-Alpha's APPLY/ONLY) then DO_FLAG_EVAL_ONLY should be
+// used--otherwise they will be evaluated normally.
+//
+// NOTE: Ren-C no longer supports the built-in ability to supply refinements
+// positionally, due to the brittleness of this approach (for both system
+// and user code).  The `opt_head` value should be made a path with the
+// function at the head and the refinements specified there.  Future
+// additions could do this more efficiently by allowing the refinement words
+// to be pushed directly to the data stack.
+//
+// !!! C's va_lists are very dangerous, there is no type checking!  The
+// C++ build should be able to check this for the callers of this function
+// *and* check that you ended properly.  It means this function will need
+// two different signatures (and so will each caller of this routine).
+//
+// Returns THROWN_FLAG, END_FLAG, or VA_LIST_FLAG
+//
+inline static REBIXO Do_Va_Core(
+    REBVAL *out,
+    const REBVAL *opt_first,
+    va_list *vaptr,
+    REBFLGS flags
+) {
+    REBFRM f;
+
+    if (opt_first)
+        SET_FRAME_VALUE(&f, opt_first); // doesn't need specifier, not relative
+    else {
+        SET_FRAME_VALUE(&f, va_arg(*vaptr, const REBVAL*));
+        assert(!IS_RELATIVE(f.value));
+    }
+
+    if (IS_END(f.value)) {
+        SET_VOID(out);
+        return END_FLAG;
+    }
+
+    f.out = out;
+#if !defined(NDEBUG)
+    f.index = TRASHED_INDEX;
+#endif
+    f.source.vaptr = vaptr;
+    f.gotten = NULL; // so ET_WORD and ET_GET_WORD do their own Get_Var
+    f.specifier = SPECIFIED; // va_list values MUST be full REBVAL* already
+    f.pending = VA_LIST_PENDING;
+
+    f.flags = flags | DO_FLAG_VA_LIST; // see notes in %sys-do.h on why needed
+
+    f.eval_type = Eval_Table[VAL_TYPE(f.value)];
+
+    Do_Core(&f);
+
+    if (THROWN(f.out))
+        return THROWN_FLAG; // !!! prohibits recovery from exits
+
+    return IS_END(f.value) ? END_FLAG : VA_LIST_FLAG;
+}
+
+
+// Wrapper around Do_Va_Core which has the actual variadic interface (as
+// opposed to taking the `va_list` whicih has been captured out of the
+// variadic interface).
+//
+inline static REBOOL Do_Va_Throws(REBVAL *out, ...)
+{
+    va_list va;
+    va_start(va, out); // must mention last param before the "..."
+
+#ifdef VA_END_IS_MANDATORY
+    struct Reb_State state;
+    REBCTX *error;
+
+    PUSH_TRAP(&error, &state);
+
+// The first time through the following code 'error' will be NULL, but...
+// `fail` can longjmp here, so 'error' won't be NULL *if* that happens!
+
+    if (error) {
+        va_end(va); // interject cleanup of whatever va_start() set up...
+        fail (error); // ...then just retrigger error
+    }
+#endif
+
+    REBIXO indexor = Do_Va_Core(
+        out,
+        NULL, // opt_first
+        &va,
+        DO_FLAG_TO_END | DO_FLAG_ARGS_EVALUATE | DO_FLAG_LOOKAHEAD
+    );
+
+    va_end(va);
+    //
+    // ^-- This va_end() will *not* be called if a fail() happens to longjmp
+    // during the apply.  But is that a problem, you ask?  No survey has
+    // turned up an existing C compiler where va_end() isn't a NOOP.
+    //
+    // But there's implementations we know of, then there's the Standard...
+    //
+    //    http://stackoverflow.com/a/32259710/211160
+    //
+    // The Standard is explicit: an implementation *could* require calling
+    // va_end() if it wished--it's undefined behavior if you skip it.
+    //
+    // In the interests of efficiency and not needing to set up trapping on
+    // each apply, our default is to assume the implementation does not
+    // need the va_end() call.  But for thoroughness, VA_END_IS_MANDATORY is
+    // outlined here to show the proper bracketing if it were ever needed.
+
+#ifdef VA_END_IS_MANDATORY
+    DROP_TRAP_SAME_STACKLEVEL_AS_PUSH(&state);
+#endif
+
+    assert(indexor == THROWN_FLAG || indexor == END_FLAG);
+    return LOGICAL(indexor == THROWN_FLAG);
+}
+
+
+// Gets a system function with tolerance of it not being a function.
+//
+// (Extraction of a feature that formerly was part of a dedicated dual
+// function to Apply_Func_Throws (Do_Sys_Func_Throws())
+//
+inline static REBVAL *Sys_Func(REBCNT inum)
+{
+    REBVAL *value = CTX_VAR(Sys_Context, inum);
+
+    if (!IS_FUNCTION(value)) fail (Error(RE_BAD_SYS_FUNC, value));
+
+    return value;
+}
+
+
+// Takes a list of arguments terminated by END_CELL (or any IS_END) and
+// will do something similar to R3-Alpha's "apply/only" with a value.  If
+// that value is a function, it will be called...if it is a SET-WORD! it
+// will be assigned, etc.
+//
+// This is equivalent to putting the value at the head of the input and
+// then calling EVAL/ONLY on it.  If all the inputs are not consumed, an
+// error will be thrown.
+//
+// The boolean result will be TRUE if an argument eval or the call created
+// a THROWN() value, with the thrown value in `out`.
+//
+inline static REBOOL Apply_Only_Throws(
+    REBVAL *out,
+    const REBVAL *applicand,
+    ...
+) {
+    va_list va;
+    va_start(va, applicand); // must mention last param before the "..."
+
+#ifdef VA_END_IS_MANDATORY
+    struct Reb_State state;
+    REBCTX *error;
+
+    PUSH_TRAP(&error, &state);
+
+// The first time through the following code 'error' will be NULL, but...
+// `fail` can longjmp here, so 'error' won't be NULL *if* that happens!
+
+    if (error) {
+        va_end(va); // interject cleanup of whatever va_start() set up...
+        fail (error); // ...then just retrigger error
+    }
+#endif
+
+    REBIXO indexor = Do_Va_Core(
+        out,
+        applicand, // opt_first
+        &va,
+        DO_FLAG_NEXT | DO_FLAG_NO_ARGS_EVALUATE | DO_FLAG_LOOKAHEAD
+    );
+
+    if (indexor == VA_LIST_FLAG) {
+        //
+        // Not consuming all the arguments given suggests a problem as far
+        // as this interface is concerned.  To tolerate incomplete states,
+        // use Do_Va_Core() directly.
+        //
+        fail (Error(RE_APPLY_TOO_MANY));
+    }
+
+    va_end(va); // see notes in Do_Va_Core RE: VA_END_IS_MANDATORY
+
+#ifdef VA_END_IS_MANDATORY
+    DROP_TRAP_SAME_STACKLEVEL_AS_PUSH(&state);
+#endif
+
+    assert(indexor == THROWN_FLAG || indexor == END_FLAG);
+    return LOGICAL(indexor == THROWN_FLAG);
+}
+
 
 inline static REBOOL Do_At_Throws(
     REBVAL *out,
@@ -629,305 +960,3 @@ inline static REBOOL EVAL_VALUE_CORE_THROWS(
 
 #define EVAL_VALUE_THROWS(out,value) \
     EVAL_VALUE_CORE_THROWS((out), (value), SPECIFIED)
-
-
-//=////////////////////////////////////////////////////////////////////////=//
-//
-//  PATH VALUE STATE "PVS"
-//
-//=////////////////////////////////////////////////////////////////////////=//
-//
-// The path value state structure is used by `Do_Path_Throws()` and passed
-// to the dispatch routines.  See additional comments in %c-path.c.
-//
-
-struct Reb_Path_Value_State {
-    //
-    // `item` is the current element within the path that is being processed.
-    // It is advanced as the path is consumed.
-    //
-    const RELVAL *item;
-
-    // A specifier is needed because the PATH! is processed by incrementing
-    // through values, which may be resident in an array that was part of
-    // the cloning of a function body.  The specifier allows the path
-    // evaluation to disambiguate which variable a word's relative binding
-    // would match.
-    //
-    REBCTX *item_specifier;
-
-    // `selector` is the result of evaluating the current path item if
-    // necessary.  So if the path is `a/(1 + 2)` and processing the second
-    // `item`, then the selector would be the computed value `3`.
-    //
-    // (This is what the individual path dispatchers should use.)
-    //
-    const REBVAL *selector;
-
-    // !!! `selector_temp` was added as a patch to push the temporary
-    // variable used to hold evaluated selectors into the PVS, when it was
-    // observed that callers of Next_Path() were expecting the selector to
-    // survive the call.  That meant it couldn't be in a C stack temporary.
-    // The method needs serious review, but keeping the temporary used for
-    // the selector in the PVS doesn't incur any more storage and bridges
-    // past the problem for now.
-    //
-    REBVAL selector_temp;
-
-    // `value` holds the path value that should be chained from.  (It is the
-    // type of `value` that dictates which dispatcher is given the `selector`
-    // to get the next step.)  This has to be a relative value in order to
-    // use the SET_IF_END option which writes into arrays.
-    //
-    RELVAL *value;
-
-    // `value_specifier` has to be updated whenever value is updated
-    //
-    REBCTX *value_specifier;
-
-    // `store` is the storage for constructed values, and also where any
-    // thrown value will be written.
-    //
-    REBVAL *store;
-
-    // `setval` is non-NULL if this is a SET-PATH!, and it is the value to
-    // ultimately set the path to.  The set should only occur at the end
-    // of the path, so most setters should check `IS_END(pvs->item + 1)`
-    // before setting.
-    //
-    // !!! See notes in %c-path.c about why the path dispatch is more
-    // complicated than simply being able to only pass the setval to the last
-    // item being dispatched (which would be cleaner, but some cases must
-    // look ahead with alternate handling).
-    //
-    const REBVAL *opt_setval;
-
-    // `orig` original path input, saved for error messages
-    //
-    const RELVAL *orig;
-};
-
-
-enum Path_Eval_Result {
-    PE_OK, // pvs->value points to the element to take the next selector
-    PE_SET_IF_END, // only sets if end of path
-    PE_USE_STORE, // set pvs->value to be pvs->store
-    PE_NONE // set pvs->store to NONE and then pvs->value to pvs->store
-};
-
-typedef REBINT (*REBPEF)(REBPVS *pvs); // Path evaluator function
-
-typedef REBINT (*REBCTF)(const RELVAL *a, const RELVAL *b, REBINT s);
-
-
-//=////////////////////////////////////////////////////////////////////////=//
-//
-//  ARGUMENT AND PARAMETER ACCESS HELPERS
-//
-//=////////////////////////////////////////////////////////////////////////=//
-//
-// These accessors are designed to make it convenient for natives and actions
-// written in C to access their arguments and refinements.  They are able
-// to bind to the implicit Reb_Frame* passed to every REBNATIVE() and read
-// the information out cleanly, like this:
-//
-//     PARAM(1, foo);
-//     REFINE(2, bar);
-//
-//     if (IS_INTEGER(ARG(foo)) && REF(bar)) { ... }
-//
-// Under the hood `PARAM(1, foo)` and `REFINE(2, bar)` make const structs.
-// In an optimized build, these structures disappear completely, with all
-// addressing done directly into the call frame's cached `arg` pointer.
-// It is also possible to get the typeset-with-symbol for a particular
-// parameter or refinement, e.g. with `PAR(foo)` or `PAR(bar)`.
-//
-// The PARAM and REFINE macros use token pasting to name the variables they
-// are declaring `p_name` instead of just `name`.  This prevents collisions
-// with C++ identifiers, so PARAM(case) and REFINE(new) would make `p_case`
-// and `p_new` instead of just `case` and `new` as the variable names.  (This
-// is only visible in the debugger.)
-//
-// As a further aid, the debug build version of the structures contain the
-// actual pointers to the arguments.  It also keeps a copy of a cache of the
-// type for the arguments, because the numeric type encoding in the bits of
-// the header requires a debug call (or by-hand-binary decoding) to interpret
-// Whether a refinement was used or not at time of call is also cached.
-//
-
-struct Native_Param {
-#if !defined(NDEBUG)
-    enum Reb_Kind kind_cache;
-    REBVAL *arg;
-#endif
-    int num;
-};
-
-struct Native_Refine {
-#if !defined(NDEBUG)
-    REBOOL used_cache;
-    REBVAL *arg;
-#endif
-    int num;
-};
-
-#ifdef NDEBUG
-    #define PARAM(n,name) \
-        const struct Native_Param p_##name = {n}
-
-    #define REFINE(n,name) \
-        const struct Native_Refine p_##name = {n}
-#else
-    // Capture the argument (and its type) for debug inspection.
-    //
-    #define PARAM(n,name) \
-        const struct Native_Param p_##name = { \
-            VAL_TYPE(frame_->arg + (n) - 1), \
-            frame_->arg + (n) - 1, \
-            (n) \
-        }
-
-    // As above, do a cache and be tolerant of framelessness.
-    //
-    #define REFINE(n,name) \
-        const struct Native_Refine p_##name = { \
-            IS_CONDITIONAL_TRUE(frame_->arg + (n) - 1), \
-            frame_->arg + (n) - 1, \
-            (n) \
-        }
-#endif
-
-// Though REF can only be used with a REFINE() declaration, ARG can be used
-// with either.
-//
-#define ARG(name) \
-    (frame_->arg + (p_##name).num - 1)
-
-#define PAR(name) \
-    FUNC_PARAM(frame_->func, (p_##name).num) // a TYPESET!
-
-#ifdef NDEBUG
-    #define REF(name) \
-        IS_CONDITIONAL_TRUE(ARG(name))
-#else
-    // An added useless ?: helps check in debug build to make sure we do not
-    // try to use REF() on something defined as PARAM(), but only REFINE()
-    //
-    #define REF(name) \
-        ((p_##name).used_cache \
-            ? IS_CONDITIONAL_TRUE(ARG(name)) \
-            : IS_CONDITIONAL_TRUE(ARG(name)))
-#endif
-
-
-//=////////////////////////////////////////////////////////////////////////=//
-//
-//  CALL FRAME ACCESS
-//
-//=////////////////////////////////////////////////////////////////////////=//
-//
-// !!! To be documented and reviewed.  Legacy naming conventions when the
-// arguments to functions lived in the data stack gave the name "FS_TOP" for
-// "(D)ata (S)tack (F)rame" which is no longer accurate, as well as the
-// convention of prefix with a D_.  The new PARAM()/REFINE()/ARG()/REF()
-// scheme is coming up on replacing a lot of it, so these will be needing
-// a tune up and choices of better names once that is sorted out.  It
-// may be as simple as changing these to FS/FSP for "frame stack pointer"
-//
-
-#define FS_TOP (TG_Frame_Stack + 0) // avoid assignment to FS_TOP via + 0
-
-#define FRM_IS_VALIST(f) \
-    LOGICAL((f)->flags & DO_FLAG_VA_LIST)
-
-#define FRM_ARRAY(f) \
-    (assert(!FRM_IS_VALIST(f)), (f)->source.array)
-
-// !!! Though the evaluator saves its `indexor`, the index is not meaningful
-// in a valist.  Also, if `opt_head` values are used to prefetch before an
-// array, those will be lost too.  A true debugging mode would need to
-// convert these cases to ordinary arrays before running them, in order
-// to accurately present the errors.
-//
-#define FRM_INDEX(f) \
-    (assert(!FRM_IS_VALIST(f)), IS_END((f)->value) \
-        ? ARR_LEN((f)->source.array) : (f)->index - 1)
-
-#define FRM_EXPR_INDEX(f) \
-    (assert(!FRM_IS_VALIST(f)), (f)->expr_index == END_FLAG \
-        ? ARR_LEN((f)->source.array) : (f)->expr_index - 1)
-
-#define FRM_OUT(f)          cast(REBVAL * const, (f)->out) // writable Lvalue
-#define FRM_PRIOR(f)        ((f)->prior)
-#define FRM_LABEL(f)        ((f)->label)
-
-#define FRM_FUNC(f)         ((f)->func)
-#define FRM_DSP_ORIG(f)     ((f)->dsp_orig + 0) // Lvalue
-
-// `arg` is in use to point at the arguments during evaluation, and `param`
-// may hold a SET-WORD! or SET-PATH! available for a lookback to quote.
-// But during evaluations, `refine` is free.
-//
-// Since the GC is aware of the pointers, it can protect whatever refine is
-// pointing at.  This can be useful for routines that have a local
-// memory cell.  This does not require a push or a pop of anything--it only
-// protects as long as the native is running.  (This trick is available to
-// the dispatchers as well.)
-//
-#define PROTECT_FRM_X(f,v) \
-    ((f)->refine = (v))
-
-// It's not clear exactly in which situations one might be using this; while
-// it seems that when filling function args you could just assume it hasn't
-// been reified, there may be "pre-reification" in the future, and also a
-// tail call optimization or some other "reuser" of a frame may jump in and
-// reuse a frame that's been reified after its initial "chunk only" state.
-// For now check the flag and don't just assume it's a raw frame.
-//
-// Uses ARR_AT instead of CTX_VAR because the varlist may not be finished.
-//
-#define FRM_ARGS_HEAD(f) \
-    ((f)->stackvars ? &(f)->stackvars[0] : KNOWN(ARR_AT((f)->varlist, 1)))
-
-// ARGS is the parameters and refinements
-// 1-based indexing into the arglist (0 slot is for object/function value)
-#ifdef NDEBUG
-    #define FRM_ARG(f,n)    ((f)->arg + (n) - 1)
-#else
-    #define FRM_ARG(f,n)    FRM_ARG_Debug((f), (n)) // checks arg index bound
-#endif
-
-// Note about D_NUM_ARGS: A native should generally not detect the arity it
-// was invoked with, (and it doesn't make sense as most implementations get
-// the full list of arguments and refinements).  However, ACTION! dispatch
-// has several different argument counts piping through a switch, and often
-// "cheats" by using the arity instead of being conditional on which action
-// ID ran.  Consider when reviewing the future of ACTION!.
-//
-#define FRM_NUM_ARGS(f) \
-    cast(REBCNT, FUNC_NUM_PARAMS((f)->func))
-
-// Quick access functions from natives (or compatible functions that name a
-// Reb_Frame pointer `frame_`) to get some of the common public fields.
-//
-#define D_OUT       FRM_OUT(frame_)         // GC-safe slot for output value
-#define D_ARGC      FRM_NUM_ARGS(frame_)        // count of args+refinements/args
-#define D_ARG(n)    FRM_ARG(frame_, (n))    // pass 1 for first arg
-#define D_REF(n)    IS_CONDITIONAL_TRUE(D_ARG(n))  // REFinement (!REFerence)
-#define D_FUNC      FRM_FUNC(frame_)        // REBVAL* of running function
-#define D_LABEL_SYM FRM_LABEL(frame_)       // symbol or placeholder for call
-#define D_DSP_ORIG  FRM_DSP_ORIG(frame_)    // Original data stack pointer
-
-#define D_PROTECT_X(v)      PROTECT_FRM_X(frame_, (v))
-#define D_PROTECT_Y(v)      PROTECT_FRM_Y(frame_, (v))
-
-// Frameless native access
-//
-// !!! Should `frame_` just be renamed to `c_` to make this briefer and be used
-// directly?  It is helpful to have macros to find the usages, however.
-//
-#define D_FRAME      frame_
-#define D_ARRAY     (frame_->source.array)
-#define D_INDEXOR   (frame_->indexor)
-#define D_VALUE     (frame_->value)
-#define D_MODE      (frame_->mode)
