+++ src/include/sys-rebser.h
@@ -425,7 +425,7 @@ struct Reb_Series {
         } area;
         REBOOL negated; // for bitsets (must be shared, can't be in REBVAL)
         REBFUN *underlying; // specialization -or- final underlying function
-        struct Reb_Frame *f; // for a FRAME! series, the call frame (or NULL)
+        REBFRM *f; // for a FRAME! series, the call frame (or NULL)
         void *fd; // file descriptor for library
         REBSTR *canon; // canon cased form of this symbol (if not canon)
         struct {
